Report data usage in `mc stat` of bucket
This PR also deprecates `mc admin bucket info` and integrates
the bucket usage metrics into `mc stat <alias>/<bucket>`
